refactor(financeiro): split PosicaoAgora (502→94) em hook + lógica + componentes
Extrai verbatim de src/components/financeiro/cashflow/PosicaoAgora.tsx:
- usePosicaoAgora: carrega capital de giro (getCapitalDeGiro) + posição ativa
- consolidate.ts (buildConsolidated — médias ponderadas PMR/PMP por volume),
  stress.ts (SCENARIOS + computeStressRow — sensibilidade do caixa)
- KpiCards, CicloFinanceiroCard, Projecao30dCard, ComparativoEmpresasCard,
  ConcentracaoRiscoCard, StressTest, MetricCard, WaterfallBar, format.ts
- +12 testes (lógica pura de consolidação/stress + format + presentacionais)

Comportamento preservado 1:1. tsc baseline=0, eslint=0 erros
(1 warning exhaustive-deps pré-existente, relocado pro hook).
